+++ docker/docker-compose.override.yml
@@ -42,7 +42,8 @@ services:
     image: locksmith
     build:
       context: ../
-      dockerfile: ./docker/locksmith.dockerfile
+      args:
+        BUILD_DIR: locksmith
       cache_from:
         - locksmith
         - unlockprotocol/locksmith:master
@@ -56,7 +57,8 @@ services:
     image: locksmith
     build:
       context: ../
-      dockerfile: ./docker/locksmith.dockerfile
+      args:
+        BUILD_DIR: locksmith
       cache_from:
         - locksmith
         - unlockprotocol/locksmith:master
@@ -75,7 +77,11 @@ services:
       - graph-node
 
   wedlocks:
-    image: unlockprotocol/wedlocks
+    image: wedlocks
+    build:
+      context: ../
+      args:
+          BUILD_DIR: locksmith
     command: yarn dev
     env_file: development.env
     ports:
@@ -92,7 +98,8 @@ services:
     env_file: development.env
     build:
       context: ../
-      dockerfile: ./docker/nudge.dockerfile
+      args:
+          BUILD_DIR: nudge
       cache_from:
         - nudge
         - unlockprotocol/nudge:master
@@ -107,7 +114,8 @@ services:
     env_file: development.env
     build:
       context: ../
-      dockerfile: ./docker/nudge.dockerfile
+      args:
+          BUILD_DIR: nudge
       cache_from:
         - nudge
         - unlockprotocol/nudge:master
